Added init and destroy methods
Added initMethod in person that is called via the beans.xml upon creation of the bean. This is set within the init value of the bean.
Added onDestroy method - called only on singletons and within the on bean.xml

default init and destroy methods can be chosen, but the methods must match the variable placed into the bean init and destroy area. So that the correct method can be found.